BLUE ATT@CK QUIZ 5 Answer (#135)
* Update blue_4.py

According to https://attack.mitre.org/techniques/T1544/ it's not T1105 (Ingress File Transfer). What about T1570?

* Update blue_4.py

in the Entreprise layer, Lateral tool transfer (T1570) is closer to the procedure suggested and T1544 (in the comment in the code) is only visible in the mobile layer; and as the procedure suggests the windows sandcat agent, I first did not thin to change to the mobile layer to find the remote copy technique.